move ExpressionOptimizer into codegen package
